iLike

+++ deploy/data/initdb.sql
@@ -5,7 +5,7 @@
 -- Dumped from database version 12.5 (Ubuntu 12.5-0ubuntu0.20.04.1)
 -- Dumped by pg_dump version 13.1
 
--- Started on 2021-02-03 18:42:02 +08
+-- Started on 2021-02-03 18:59:20 +08
 
 SET statement_timeout = 0;
 SET lock_timeout = 0;
@@ -157,7 +157,7 @@ CREATE TABLE public.item (
     img_url character varying(350),
     content character varying(2000),
     type_id integer DEFAULT 0,
-    price money DEFAULT 0,
+    price double precision DEFAULT 0,
     quantity integer DEFAULT 0,
     status integer DEFAULT 0
 );
@@ -410,7 +410,7 @@ CREATE TABLE public.subscription (
     img_url character varying(350),
     content character varying(2000),
     type_id integer,
-    price money DEFAULT 0,
+    price double precision DEFAULT 0,
     quantity integer DEFAULT 0,
     status integer
 );
@@ -1149,7 +1149,7 @@ CREATE INDEX tournament_lower_idx ON public.tournament USING btree (lower((title
 CREATE INDEX tournament_set_lower_idx ON public.tournament_set USING btree (lower((title)::text));
 
 
--- Completed on 2021-02-03 18:42:07 +08
+-- Completed on 2021-02-03 18:59:24 +08
 
 --
 -- PostgreSQL database dump complete
